pyproject.toml
[tool.poetry]
name = "procollab-backend"
version = "0.1.0"
description = ""
authors = ["Sergey Yaksanov <sergeyyaksanov@yandex.ru>"]
readme = "README.md"
packages = []

[tool.black]
    line-length = 90
    include = '\.pyi?$'
    exclude = '''
    /(
        \.git
      | \.hg
      | \.mypy_cache
      | \.tox
      | \.venv
      | _build
      | buck-out
      | build
      | dist
      | venv
      | migrations
      | blib2to3
      | tests/data
    )/
    '''

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
flake8 = "^5.0.4"
python-decouple = "^3.6"
psycopg2-binary = "^2.9.4"
django-cleanup = "^6.0.0"
django-cors-headers = "^3.13.0"
django-debug-toolbar = "^3.7.0"
Pillow = "^9.2.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.1"}
pre-commit = "^2.20.0"
black = "^22.10.0"
uritemplate = "^4.1.1"
PyYAML = "^6.0"
django-rest-passwordreset = "^1.3.0"
django-filter = "^22.1"
setuptools = "^65.5.0"
drf-yasg = "^1.21.4"
sentry-sdk = "^1.10.1"
whitenoise = "^6.2.0"
six = "^1.16.0"
aiohttp = "^3.8.3"
django = {extras = ["bcrypt"], version = "^4.1.3"}
channels-redis = "^4.0.0"
channels = {extras = ["daphne"], version = "^4.0.0"}
django-taggit = "^3.1.0"
loguru = "^0.7.0"
requests = "^2.31.0"
coreapi = "^2.3.3"
webp = "^0.1.6"
django-prometheus = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"